requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
bottlenose==1.1.8
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
numpy==1.15.4
pandas==0.23.4
py==1.7.0
pylint==2.1.1
python-dateutil==2.7.5
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
retry==0.9.2
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
